ESQL: Drop unused Node subclasses
And merge NodeSubclassTests into EsqlNodeSubclassTests.
