--- x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/plan/logical/Project.java
@@ -1,83 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-package org.elasticsearch.xpack.esql.core.plan.logical;
-
-import org.elasticsearch.xpack.esql.core.capabilities.Resolvables;
-import org.elasticsearch.xpack.esql.core.expression.Attribute;
-import org.elasticsearch.xpack.esql.core.expression.Expressions;
-import org.elasticsearch.xpack.esql.core.expression.NamedExpression;
-import org.elasticsearch.xpack.esql.core.expression.function.Functions;
-import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
-import org.elasticsearch.xpack.esql.core.tree.Source;
-
-import java.util.List;
-import java.util.Objects;
-
-/**
- * A {@code Project} is a {@code Plan} with one child. In {@code SELECT x FROM y}, the "SELECT" statement is a Project.
- */
-public class Project extends UnaryPlan {
-
-    private final List<? extends NamedExpression> projections;
-
-    public Project(Source source, LogicalPlan child, List<? extends NamedExpression> projections) {
-        super(source, child);
-        this.projections = projections;
-    }
-
-    @Override
-    protected NodeInfo<Project> info() {
-        return NodeInfo.create(this, Project::new, child(), projections);
-    }
-
-    @Override
-    public Project replaceChild(LogicalPlan newChild) {
-        return new Project(source(), newChild, projections);
-    }
-
-    public List<? extends NamedExpression> projections() {
-        return projections;
-    }
-
-    public Project withProjections(List<? extends NamedExpression> projections) {
-        return new Project(source(), child(), projections);
-    }
-
-    @Override
-    public boolean resolved() {
-        return super.resolved() && Expressions.anyMatch(projections, Functions::isAggregate) == false;
-    }
-
-    @Override
-    public boolean expressionsResolved() {
-        return Resolvables.resolved(projections);
-    }
-
-    @Override
-    public List<Attribute> output() {
-        return Expressions.asAttributes(projections);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(projections, child());
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null || getClass() != obj.getClass()) {
-            return false;
-        }
-
-        Project other = (Project) obj;
-
-        return Objects.equals(projections, other.projections) && Objects.equals(child(), other.child());
-    }
-}
